Add Tripal/Chado FormatterBases classes.
