+++ tripal_chado/src/Plugin/Field/FieldFormatter/OBIOrganismDefaultFormatter.php
@@ -5,7 +5,7 @@ namespace Drupal\tripal_chado\Plugin\Field\FieldFormatter;
 use Drupal\Component\Utility\Html;
 use Drupal\Core\Field\FieldItemInterface;
 use Drupal\Core\Field\FieldItemListInterface;
-use Drupal\Core\Field\FormatterBase;
+use Drupal\tripal_chado\Plugin\Field\ChadoFormatterBase;
 use Drupal\Core\Form\FormStateInterface;
 
 /**
@@ -19,14 +19,25 @@ use Drupal\Core\Form\FormStateInterface;
  *   }
  * )
  */
-class OBIOrganismDefaultFormatter extends FormatterBase {
+class OBIOrganismDefaultFormatter extends ChadoFormatterBase {
+
+  /**
+   * @param array
+   *   The options available for format.
+   *   These map to the keys of the field property definition.
+   */
+  public static $format_options = [
+    'scientific_name' => 'Scientific Name',
+    'common_name' => 'Common Name',
+    'abbreviation' => 'Abbreviation',
+  ];
 
   /**
    * {@inheritdoc}
    */
   public static function defaultSettings() {
     return [
-      // Implement default settings.
+      'format' => 'scientific_name',
     ] + parent::defaultSettings();
   }
 
@@ -34,60 +45,31 @@ class OBIOrganismDefaultFormatter extends FormatterBase {
    * {@inheritdoc}
    */
   public function settingsForm(array $form, FormStateInterface $form_state) {
-    return [
-      // Implement settings form.
-    ] + parent::settingsForm($form, $form_state);
+    $element = [];
+
+    $element['format'] = [
+      '#type' => 'select',
+      '#title' => t('Format'),
+      '#description' => t('The format to display the organism using.'),
+      '#options' => $this::$format_options,
+      '#default_value' => $this->getSetting('format'),
+    ];
+
+    $element += parent::settingsForm($form, $form_state);
+    return $element;
   }
 
   /**
    * {@inheritdoc}
    */
   public function settingsSummary() {
-    $summary = [];
-    // Implement settings summary.
 
+    $format = $this->getSetting('format');
+    $options = $this::$format_options;
+    $summary = [
+      '#markup' => 'Format: ' . $options[$format],
+    ];
     return $summary;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function viewElements(FieldItemListInterface $items, $langcode) {
-    $elements = [];
-
-    foreach ($items as $delta => $item) {
-      $elements[$delta] = ['#markup' => $this->viewValue($item)];
-    }
-
-    return $elements;
-  }
-
-  /**
-   * Retrieve a specific value from the items list.
-   *
-   * @todo move this into a ChadoWidgetBase class.
-   *
-   * @param array $item
-   *   The values for the OBIOrganismItem field on this entity.
-   * @param string $property_name
-   *   The name of the value or property you would like to pull out. Supported
-   *   values include record_id, chado_schema, etc.
-   */
-  public function getChadoValue($item, $property_name) {
-
-    if ($property_name == 'record_id') {
-      return $item->get('record_id')->getValue();
-    }
-    elseif ($property_name == 'chado_schema') {
-      return $item->get('chado_schema')->getValue();
-    }
-    else {
-      $values = unserialize($item->getValue());
-      if (isset($values[$property_name])) {
-        return $values[$property_name];
-      }
-    }
-    return NULL;
   }
 
   /**
@@ -101,12 +83,8 @@ class OBIOrganismDefaultFormatter extends FormatterBase {
    */
   protected function viewValue(FieldItemInterface $item) {
 
-    $genus = $this->getChadoValue($item, 'genus');
-    $species = $this->getChadoValue($item, 'species');
-    $common_name = $this->getChadoValue($item, 'common_name');
-
-    $output = $genus . ' ' . $species;
-
+    $format = $this->getSetting('format');
+    $output = $this->getChadoValue($item, $format);
     return nl2br(Html::escape($output));
   }
